requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
langchain-community==0.0.28
tiktoken==0.6.0
unstructured==0.12.6
python-dotenv==1.0.1
chromadb==0.4.24